+++ lib/src/extensions/colors_ext.dart
@@ -1,0 +1,25 @@
+import 'package:flutter/material.dart';
+
+extension HexColor on Color {
+  /// Prefixes a hash sign if [leadingHashSign] is set to `true` (default is `true`).
+  String toHex({bool leadingHashSign = true}) => '${leadingHashSign ? '#' : ''}'
+      '${alpha.toRadixString(16).padLeft(2, '0')}'
+      '${red.toRadixString(16).padLeft(2, '0')}'
+      '${green.toRadixString(16).padLeft(2, '0')}'
+      '${blue.toRadixString(16).padLeft(2, '0')}';
+}
+
+extension ColorExt on String {
+  bool get isHexColor =>
+      RegExp(r'^#?(?:[0-9a-fA-F]{3,4}){1,2}$').hasMatch(this);
+
+  Color? get toColor {
+    if (isHexColor) {
+      final buffer = StringBuffer();
+      if (length == 6 || length == 7) buffer.write('FF');
+      buffer.write(replaceFirst('#', ''));
+      return Color(int.parse(buffer.toString(), radix: 16));
+    }
+    return null;
+  }
+}