+++ lib/src/extensions/extensions.dart
@@ -1,3 +1,4 @@
+export 'colors_ext.dart';
 export 'data_structures/stack.dart';
 export 'date.dart';
 export 'duration.dart';